refactor(setup): extract stepMemoryBackend so its recall-pool test runs under vitest
setup-recall-pool-provision.test.ts is a vitest test (imports from "vitest",
not bun:test) but imported stepMemoryBackend from ./setup.js, whose static
import graph reaches the vault → bun:sqlite edge (../vault/vault.js,
./vault-broker.js, ../vault/broker/client.js as a value, ./setup-posture-rewrite.js).
vitest cannot load bun:sqlite, so the test was excluded from vitest AND named in
test:bun — but the bun CI runner runs from telegram-plugin/ with plugin-relative
paths, so a repo-root src/cli/... file was executed by NEITHER runner. That is
exactly the "file in no runner" state check-test-runner-coverage exists to catch.

Fix (durable — make the test actually RUN):
- Extract stepMemoryBackend + MemoryBackendDeps/MemoryBackendOutcome/
  isMemoryBackendDisabled + the READY constants into a new vault-free sibling
  src/cli/setup-memory-backend.ts. setup.ts imports stepMemoryBackend from it
  (call site at step 6 unchanged); the test imports from it.
- The one legacy vault touchpoint inside the step (getStringSecret for the
  deprecated hindsight-api-key courtesy note) is now reached via a dynamic
  `await import("../vault/vault.js")` inside its try block, so the module's
  STATIC graph stays free of bun:sqlite while runtime behaviour is unchanged.
  ../vault/broker/client.js is type-only (erased). resolvePath comes from the
  vault-free ../config/loader.js.
- Shared step-header helpers (STEP_ACTIVE/STEP_DONE/stepHeader/stepHeaderTo)
  moved to src/cli/setup-step-ui.ts so both setup.ts and the new module import
  them without an import cycle.
- Remove the vitest exclude entry and drop the file from test:bun (it imports
  "vitest", so bun test would fail on it).

setup-recall-pool-provision.test.ts now loads and passes under vitest (3 tests).

Switchroom-Agent: klanker
Switchroom-Model: claude-opus-4-8
